Add `workerd` and `unwasm` condition
