+++ js/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "md4w",
+  "description": "A WebAssembly port of md4c - a Markdown parser written in C.",
+  "version": "0.0.2",
+  "type": "module",
+  "module": "index.js",
+  "main": "index.js",
+  "exports": {
+    ".": {
+      "workerd": "./unwasm.js",
+      "unwasm": "./unwasm.js",
+      "import": "./index.js"
+    }
+  },
+  "license": "MIT"
+}
